Add and apply GameUtil
